Setup Wizard: Site Admin Repo Sync view (#47141)
Closes #47110
([Figma](https://www.figma.com/file/EuRdjjQF8a9kRJMRce4HyI/UX%3A-Instance-setup-experience?node-id=746%3A13794&t=oLVPzFL3CY2X4dkO-0)).
Reuses Site Admin Repo Sync view with a design refresh & button action
updates.

**In action clip**:

https://user-images.githubusercontent.com/59381432/216470321-cda294a0-8821-4a2f-ab17-50c72398d96c.mov
<img width="1378" alt="Screenshot 2023-02-08 at 5 09 35 PM"
src="https://user-images.githubusercontent.com/59381432/217585873-26f4e4c7-527f-4055-a32c-7f35c80a08bc.png">
<img width="1064" alt="Screenshot 2023-02-08 at 5 11 17 PM"
src="https://user-images.githubusercontent.com/59381432/217585949-6aff7078-82c1-4fad-ac37-4f6acdb2d6e6.png">


## Changelog
- Makes `<SiteAdminRepositories />` flexible to be reused in our use
case
- Implement dark/light theme color styles
- Move conditional /setup route with other routes for dark/light theme
color constant access

#### Design refresh for `<SiteAdminRepositoriesPage />`:
- Add Badge for repo status
- Rework `<RepoMirrorInfo />` because new status UI will overtake some
of this visual data
- Add loading state (no % available)

## Test plan
- Turn setup feature flag on in settings (`"experimentalFeatures": {
"setupWizard": true }`)
- Visit `/site-admin/repositories` or `/setup` to view changed
components

## Note
- Note, Footer is out of view on page. I think if we contain this info
in a smaller scrollable vertical space it will be too squished though.
Open to other thoughts!
**Follow up issues**: ~#47393~, #47394, #47395

## App preview:

-
[Web](https://sg-web-becca-setup-wiz-sync-repo-step.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.
