+++ client/web/src/components/FilteredConnection/FilterControl.module.scss
@@ -2,6 +2,6 @@
     display: flex;
     align-items: center;
     flex-wrap: wrap;
-    gap: 1rem;
+    gap: 0.5rem;
     max-width: 60%;
 }
