+++ client/web/src/components/FilteredConnection/FilterControl.tsx
@@ -96,6 +96,7 @@ export const FilterControl: React.FunctionComponent<React.PropsWithChildren<Filt
                                     onChange={event => onChange(filter, event.currentTarget.value)}
                                     value={values.get(filter.id)?.value}
                                     className="mb-0"
+                                    isCustomStyle={true}
                                 >
                                     {filter.values.map(value => (
                                         <option key={value.value} value={value.value} label={value.label} />
