--- client/web/src/site-admin/SiteAdminRepositoriesPage.module.scss
@@ -1,9 +1,0 @@
-.alert-wrapper {
-    margin-top: 1rem;
-    overflow-y: auto;
-    max-height: 25rem;
-}
-
-.alert-content {
-    white-space: break-spaces;
-}
