+++ client/web/src/site-admin/analytics/components/ValueLegendList.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/forbid-dom-props */
-import React, { useMemo } from 'react'
+import React, { useMemo, ReactNode } from 'react'
 
 import classNames from 'classnames'
 import { useLocation } from 'react-router-dom'
@@ -13,7 +13,7 @@ import styles from './index.module.scss'
 interface ValueLegendItemProps {
     color?: string
     description: string
-    value: number | string
+    value: number | string | ReactNode
     tooltip?: string
     className?: string
     filter?: { name: string; value: string }
@@ -55,7 +55,13 @@ export const ValueLegendItem: React.FunctionComponent<ValueLegendItemProps> = ({
                         {formattedNumber}
                     </Link>
                 ) : (
-                    <Text as="span" alignment="center" style={{ color }} className={styles.count} onClick={onClick}>
+                    <Text
+                        as="span"
+                        alignment="center"
+                        style={{ color }}
+                        className={classNames(styles.count, 'cursor-pointer')}
+                        onClick={onClick}
+                    >
                         {formattedNumber}
                     </Text>
                 )}
