+++ client/web/src/site-admin/components/RepoMirrorInfo.tsx
@@ -1,9 +1,7 @@
 import * as React from 'react'
 
-import { mdiCloudOutline } from '@mdi/js'
-
 import { Timestamp } from '@sourcegraph/branded/src/components/Timestamp'
-import { Icon, LoadingSpinner, Text, Tooltip } from '@sourcegraph/wildcard'
+import { Text, Tooltip } from '@sourcegraph/wildcard'
 
 import { MirrorRepositoryInfoFields } from '../../graphql-operations'
 import { prettyBytesBigint } from '../../util/prettyBytesBigint'
@@ -14,18 +12,6 @@ export const RepoMirrorInfo: React.FunctionComponent<
     }>
 > = ({ mirrorInfo }) => (
     <>
-        {mirrorInfo.cloneInProgress && (
-            <small className="ml-2 text-success">
-                <LoadingSpinner /> Cloning
-            </small>
-        )}
-        {!mirrorInfo.cloneInProgress && !mirrorInfo.cloned && (
-            <Tooltip content="Visit the repository to clone it. See its mirroring settings for diagnostics.">
-                <small className="ml-2 text-muted">
-                    <Icon aria-hidden={true} svgPath={mdiCloudOutline} /> Not yet cloned
-                </small>
-            </Tooltip>
-        )}
         <Text className="mb-0 text-muted">
             <small>
                 {mirrorInfo.updatedAt === null ? (
