+++ client/web/src/site/NeedsRepositoryConfigurationAlert.tsx
@@ -5,6 +5,7 @@ import classNames from 'classnames'
 import { Link } from '@sourcegraph/wildcard'
 
 import { DismissibleAlert } from '../components/DismissibleAlert'
+import { PageRoutes } from '../routes.constants'
 import { useExperimentalFeatures } from '../stores'
 import { eventLogger } from '../tracking/eventLogger'
 
@@ -29,7 +30,7 @@ export const NeedsRepositoryConfigurationAlert: React.FunctionComponent<
         >
             <Link
                 className="site-alert__link"
-                to={isSetupWizardEnabled ? '/setup' : '/site-admin/external-services'}
+                to={isSetupWizardEnabled ? PageRoutes.SetupWizard : '/site-admin/external-services'}
                 onClick={onClickCTA}
             >
                 <span className="underline">Connect a code host</span>
